Adicionado SweetAlert no projeto

+++ src/app/colaborador-detalhes/colaborador-detalhes.component.ts
@@ -2,6 +2,7 @@ import { ColaboradoresService } from './../services/colaboradores.service';
 import { Colaborador } from './../models/colaborador.model';
 import { Component, OnInit } from '@angular/core';
 import { Router, ActivatedRoute } from '@angular/router';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-colaborador-detalhes',
@@ -35,13 +36,45 @@ export class ColaboradorDetalhesComponent implements OnInit {
   }
 
   delete(id) {
-    this.isLoadingResults = true;
-    this.colaboradorService.delete(id).subscribe(res => {
-      this.isLoadingResults = false;
-      this.router.navigate(['colaboradores'])
-    }, (err) => {
-      console.log(err);
-      this.isLoadingResults = false;
-    });
+    const swalWithBootstrapButtons = Swal.mixin({
+      customClass: {
+        confirmButton: 'btn btn-success',
+        cancelButton: 'btn btn-danger line-spacing'
+      },
+      buttonsStyling: false
+    })
+    
+    swalWithBootstrapButtons.fire({
+      title: 'Você tem certeza que deseja excluir o colaborador?',
+      text: "Esse processo é irreversível!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Sim, pode excluir!',
+      cancelButtonText: 'Não, cancela!',
+      reverseButtons: true
+    }).then((result) => {
+      if (result.value) {
+        //realiza a exclução do colaborador
+        this.colaboradorService.delete(id).subscribe(res => {
+          this.router.navigate(['colaboradores'])
+        }, (err) => {
+          console.log(err);
+        });
+        swalWithBootstrapButtons.fire(
+          'Excluído!',
+          'Colaborador excluído com sucesso!',
+          'success'
+        )
+      } else if (
+        result.dismiss === Swal.DismissReason.cancel
+      ) {
+        swalWithBootstrapButtons.fire(
+          'Cancelado',
+          'Operação cancelada conforme solicitado :)',
+          'error'
+        )
+      }
+    })
+    
   }
 }
